Return single Polygon if type is Polygon in decodePolygon
The decodePolygon function was always returning an array of Polygons. This
would work if the type was MultiPolygon. But if there was a single Polygon
with holes, it wouldn't work because the type was Polygon but it was nested
too deep within a MultiPolygon structrue.

I changed it such that when coords.length === 1 (type is Polygon) then
return coords[0]. Otherwise it returns coords.

http://www.geojson.org/geojson-spec.html#id4